feat - Ocean biome

+++ src/world/region.rs
@@ -2,7 +2,7 @@
 pub struct Region {
     pub name: String,
     pub id: usize,
-    pub elevation: (u8, u8),
+    pub elevation: (i32, i32),
     pub temperature: (u8, u8),
     pub soil_fertility_range: (f32, f32),
     pub gold_generation_range: (f32, f32),
